Improve Firefox on iOS scroll behavior
where without this fix, when the nav bar expands/contractrs as the user manually scrolls, it changes the value of `100vh` which makes the entire page grow & shrink unpredictably, really screwing with stories near the bottom of the page.

With this new styling, Firefox on iOS still experiences some oddness if *many* stories have their transcripts open, but hopefully that's an edge case.